% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_backward_hmm.R
\name{log_backward_hmm}
\alias{log_backward_hmm}
\title{log backward probabilities of HMM}
\usage{
log_backward_hmm(K, n, h_mat, init_dist, tran_prob)
}
\arguments{
\item{K}{the number of switching states}

\item{n}{the number of time points}

\item{h_mat}{likelihood value matrix}

\item{init_dist}{initial dostribution}

\item{tran_prob}{transition probability matrix}
}
\value{
A log backward probability vector
}
\description{
Compute (log) backward probability of a hidden Markov model
}
