% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_slds.R
\name{mcmc_slds}
\alias{mcmc_slds}
\title{Main MCMC algorithm for SLDS}
\usage{
mcmc_slds(y, x_init = stats::rnorm(length(y) + 1), s_init = sample(c(1,
  2), size = length(y) + 1, replace = TRUE), v_init = 2, w_init = 2,
  phi_init = 0.5, xi_init = 0.5, av0 = 1/2, bv0 = 1/2, aw0 = 1/2,
  bw0 = 1/2, a_xi = 1, b_xi = 1, m0 = 0, C0 = 1000, k2 = 30,
  burn = 0, thin = 1, mcmc_len = 2000)
}
\arguments{
\item{y}{observation vector}

\item{x_init}{initial value of continuous latent states}

\item{s_init}{initial value of discrete hidden states}

\item{v_init}{initial value of observational variance}

\item{w_init}{initial value of evolution variance}

\item{phi_init}{initial value of AR(1) coefficient}

\item{xi_init}{initial probability of the transition matrix}

\item{av0}{shape parameter of the inverse gamma prior for v}

\item{bv0}{scale parameter of the inverse gamma prior for v}

\item{aw0}{shape parameter of the inverse gamma prior for w}

\item{bw0}{scale parameter of the inverse gamma prior for w}

\item{a_xi}{first shape parameter of the beta prior for xi}

\item{b_xi}{second shape parameter of the beta prior for xi}

\item{m0}{mean of the normal prior}

\item{C0}{(co)variance of the normal prior}

\item{k2}{observational variance when \code{s} is 2}

\item{burn}{burning period}

\item{thin}{thinning}

\item{mcmc_len}{total number of MCMC iterations}
}
\value{
A list of posterior samples of of unknown continuous and discrete latent
    states and unknown model parameters
}
\description{
Generate posterior samples of unknown continuous and discrete latent
    states as well as other unknown model parameters.
}
