% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_prob_hmm.R
\name{state_prob_hmm}
\alias{state_prob_hmm}
\title{posterior state probabilities of HMM}
\usage{
state_prob_hmm(K, n, h_mat, init_dist, tran_prob)
}
\arguments{
\item{K}{the number of switching states}

\item{n}{the number of time points}

\item{h_mat}{likelihood value matrix}

\item{init_dist}{initial dostribution}

\item{tran_prob}{transition probability matrix}
}
\value{
A posterior state probability vector
}
\description{
Compute posterior discrete state probabilities of a hidden Markov model
}
