% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufficient_stats.R
\name{sufficient_stats}
\alias{sufficient_stats}
\title{sufficient statistics compuation}
\usage{
sufficient_stats(ms, Cs, tm1, t, GG = NULL, m0 = NULL, C0 = NULL)
}
\arguments{
\item{ms}{mean of Kalman smoother}

\item{Cs}{(co)variance of Kalman smoother}

\item{tm1}{time t - 1}

\item{t}{time t}

\item{GG}{evolution matrix}

\item{m0}{mean of initial continuous latent state}

\item{C0}{mean of initial continuous latent state}
}
\value{
A sufficient statistics value
}
\description{
Compute sufficient statistics of a state space model
}
