% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_C0hat_inv.R
\name{update_C0hat_inv}
\alias{update_C0hat_inv}
\title{Update C0hat_inv}
\usage{
update_C0hat_inv(C0_inv_lst, FF_V_inv_FF_lst, Vhat_inv, FFhat,
  GG_W_inv_GG_lst, What_inv, GGhat, qs_t)
}
\arguments{
\item{C0_inv_lst}{list of \code{C_0^{-1}} with elements being value under different
value of discrete hidden states}

\item{FF_V_inv_FF_lst}{list of FF' V^-1 FF with elements being value under different
value of discrete hidden states}

\item{Vhat_inv}{variational parameter inverse of Vhat}

\item{FFhat}{variational parameter FFhat}

\item{GG_W_inv_GG_lst}{list of GG' W^-1 GG with elements being value under different
value of discrete hidden states}

\item{What_inv}{variational parameter inverse of What}

\item{GGhat}{variational parameter GGhat}

\item{qs_t}{probability vector of the discrete hidden state at time t}
}
\value{
An updated value of C0hat inverse
}
\description{
Update the variational parameter C0hat_inv
    in the variational distribution q(X) which is a state space model
}
