% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_FFhat.R
\name{update_FFhat}
\alias{update_FFhat}
\title{Update FFhat at time t}
\usage{
update_FFhat(V_inv_FF_lst, Vhat_inv, qs_t)
}
\arguments{
\item{V_inv_FF_lst}{list of \code{V^{-1}FF} with elements being value under different
value of discrete hidden states}

\item{Vhat_inv}{variational parameter inverse of Vhat}

\item{qs_t}{probability vector of the discrete hidden state at time t}
}
\value{
An updated value of FFhat at time t
}
\description{
Update the variational parameter FFhat at time t
    in the variational distribution q(X) which is a state space model
}
