% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_GGhat.R
\name{update_GGhat}
\alias{update_GGhat}
\title{Update GGhat at time t}
\usage{
update_GGhat(What_inv, W_inv_GG_lst, qs_t)
}
\arguments{
\item{What_inv}{variational parameter inverse of What}

\item{W_inv_GG_lst}{list of \code{W^{-1} GG} with elements being value under different
value of discrete hidden states}

\item{qs_t}{probability vector of the discrete hidden state at time t}
}
\value{
An updated value of GGhat at time t
}
\description{
Update the variational parameter GGhat at time t
    in the variational distribution q(X) which is a state space model
}
