% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_m0hat.R
\name{update_m0hat}
\alias{update_m0hat}
\title{Update m0hat}
\usage{
update_m0hat(C0hat_inv, C0_inv_m0_lst, qs_t)
}
\arguments{
\item{C0hat_inv}{inverse of C0hat}

\item{C0_inv_m0_lst}{list of \code{C_0^{-1}m_0} with elements being value under different
value of discrete hidden states}

\item{qs_t}{probability vector of the discrete hidden state at time t}
}
\value{
An updated value of m0hat
}
\description{
Update the variational parameter m0hat
    in the variational distribution q(X) which is a state space model
}
