% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_phi.R
\name{update_phi}
\alias{update_phi}
\title{Update AR(1) coefficient in the evolution level process}
\usage{
update_phi(xt, xm1, w)
}
\arguments{
\item{xt}{mcmc sample of \code{x_t}, the continuous latent states}

\item{xm1}{mcmc sample of \code{x_{t-1}}, the continuous latent states}

\item{w}{mcmc sample of evolutional variance}
}
\value{
A number drawn from a normal distribtion
}
\description{
Sample AR(1) coefficient from a normal distribution
    with a flat prior with a constriant that the coefficient is in \code{[-1, 1]}.
}
