% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_s.R
\name{update_s}
\alias{update_s}
\title{Update discrete hidden states using discrete version of FFBS algorithm}
\usage{
update_s(y, xt, v, k2, tran_mat)
}
\arguments{
\item{y}{observation vector}

\item{xt}{mcmc sample of \code{x_t}, the continuous latent states}

\item{v}{mcmc sample of observational variance}

\item{k2}{observational variance when \code{s} is 2}

\item{tran_mat}{transition matrix}
}
\value{
A vector of discrete hidden state
}
\description{
Sample new discrete hidden states using discrete version of
    forward filtering backward sampling (FFBS) algorithm with a given transition
    matrix.
}
