% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_v.R
\name{update_v}
\alias{update_v}
\title{Update variance of observational noise}
\usage{
update_v(av0, bv0, y, xt, st, k2)
}
\arguments{
\item{av0}{shape parameter of the inverse gamma prior}

\item{bv0}{scale parameter of the inverse gamma prior}

\item{y}{observation vector}

\item{xt}{mcmc sample of \code{x_t}, the continuous latent states}

\item{st}{mcmc sample of \code{s_t}, the discrete hidden states}

\item{k2}{observational variance when \code{s} is 2}
}
\value{
A number drawn from an inverse gamma distribtion
}
\description{
Sample new observational variance from inverse gamma distribution
    with the inverse gamma prior \code{IG(av0, bv0)}.
}
