% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Vhat_inv.R
\name{update_Vhat_inv}
\alias{update_Vhat_inv}
\title{Update inverse of Vhat at time t}
\usage{
update_Vhat_inv(V_inv_lst, qs_t)
}
\arguments{
\item{V_inv_lst}{list of inverse of V with elements being value under different
value of discrete hidden states}

\item{qs_t}{probability vector of the discrete hidden state at time t}
}
\value{
An updated value of inverse of Vhat at time t
}
\description{
Update the variational parameter Vhat inverse at time t
    in the variational distribution q(X) which is a state space model
}
