% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_w}
\alias{update_w}
\title{Update variance of evolution noise}
\usage{
update_w(aw0, bw0, xt, xm1, phi)
}
\arguments{
\item{aw0}{shape parameter of the inverse gamma prior}

\item{bw0}{scale parameter of the inverse gamma prior}

\item{xt}{mcmc sample of \code{x_t}, the continuous latent states}

\item{xm1}{mcmc sample of \code{x_{t-1}}, the continuous latent states}

\item{phi}{mcmc sample of AR(1) coefficient}
}
\value{
A number drawn from an inverse gamma distribtion
}
\description{
Sample new evolutional variance from inverse gamma distribution
    with the inverse gamma prior \code{IG(aw0, bw0)}.
}
