% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_What_inv.R
\name{update_What_inv}
\alias{update_What_inv}
\title{Update What_inv at time t}
\usage{
update_What_inv(W_inv_lst, FF_V_inv_FF_lst, Vhat_inv, FFhat,
  GG_W_inv_GG_lst = NULL, What_inv = NULL, GGhat = NULL, qs_t, t, TT)
}
\arguments{
\item{W_inv_lst}{list of \code{W^{-1}} with elements being value under different
value of discrete hidden states}

\item{FF_V_inv_FF_lst}{list of FF %*% V^{-1} FF with elements being value under different
value of discrete hidden states}

\item{Vhat_inv}{variational parameter inverse of Vhat}

\item{FFhat}{variational parameter FFhat}

\item{GG_W_inv_GG_lst}{list of GG %*% W^{-1} GG with elements being value under different
value of discrete hidden states}

\item{What_inv}{variational parameter inverse of What}

\item{GGhat}{variational parameter GGhat}

\item{qs_t}{probability vector of the discrete hidden state at time t}

\item{t}{time t}

\item{TT}{time length}
}
\value{
An updated value of What inverse at time t
}
\description{
Update the variational parameter What_inv at time t
    in the variational distribution q(X) which is a state space model
}
