% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xi.R
\name{update_xi}
\alias{update_xi}
\title{Update probability in the transition matrix}
\usage{
update_xi(a_xi, b_xi, st)
}
\arguments{
\item{a_xi}{first shape parameter of the beta prior}

\item{b_xi}{second shape parameter of the beta prior}

\item{st}{mcmc sample of \code{s_t}, the discrete hidden states}
}
\value{
A 2 by 2 transition matrix
}
\description{
Sample probability from a beta distribution
    with a \code{beta(a_xi, b_xi)} prior
}
