% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variational_em_slds.R
\name{variational_em_slds}
\alias{variational_em_slds}
\title{Main variational EM algorithm for SLDS}
\usage{
variational_em_slds(y, qs = matrix(0.5, 2, TT), k2 = 30,
  is.obs = TRUE)
}
\arguments{
\item{y}{observations}

\item{qs}{matrix of size 2 by TT with each row being the probability vector for S_t}

\item{k2}{observational variance when st = 2}

\item{is.obs}{is h_mat obseved value or likelihood value}
}
\value{
list of variational parameters of q(X) and q(S), ELBO, and q(X) and q(S)
}
\description{
Generate approximate posterior distribution of continous and discrete latent
    state variables and point estimates for model parameterslibrary("benchmarkme")
}
