% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_mean.R
\name{wt_mean}
\alias{wt_mean}
\title{Compute weighted mean}
\usage{
wt_mean(x_lst, wt)
}
\arguments{
\item{x_lst}{a list containing elements to be weighted}

\item{wt}{weight vector}
}
\value{
A weigthed mean of a given list
}
\description{
Compute weighted means from a given list and weight vector
}
